Finished Sold Out and Exact Change, had to switch from using doubles to decimals for a strange rounding issue with some numbers.

+++ VendingMachine/Helpers/CoinHelper.cs
@@ -7,11 +7,11 @@ namespace VendingMachine.Helpers
 {
 	public static class CoinHelper
 	{
-		public static double GetCoinValue(string coinName)
+		public static decimal GetCoinValue(string coinName)
 		{
 			if (IsValidCoin(coinName))
 			{
-				return (int)Enum.Parse(typeof(AcceptedCoinEnum), coinName) / 100.0;
+				return (int)Enum.Parse(typeof(AcceptedCoinEnum), coinName) / 100.0M;
 			}
 			else
 			{
